+++ agents/config/db.go
@@ -1,4 +1,4 @@
-package executor
+package config
 
 import (
 	"context"