+++ packages/synapse-interface/pages/_app.tsx
@@ -34,8 +34,10 @@ import {
   getDefaultWallets,
   connectorsForWallets,
 } from '@rainbow-me/rainbowkit'
+import { rabbyWallet } from '@rainbow-me/rainbowkit/wallets'
 import { JsonRpcProvider } from '@ethersproject/providers'
 import { jsonRpcProvider } from 'wagmi/providers/jsonRpc'
+import { publicProvider } from 'wagmi/providers/public'
 import * as CHAINS from '@constants/chains/master'
 import { SynapseProvider } from '@/utils/providers/SynapseProvider'
 import CustomToaster from '@/components/toast'
@@ -79,7 +81,8 @@ for (const chain of rawChains) {
   chainsMatured.push({
     ...chain,
     iconUrl: configChain.chainImg.src,
-    configRpc: configChain.rpc,
+    configRpc: configChain.rpcUrls.primary,
+    fallbackRpc: configChain.rpcUrls.fallback,
   })
 }
 
@@ -91,6 +94,12 @@ const { chains, publicClient, webSocketPublicClient } = configureChains(
         http: chain['configRpc'],
       }),
     }),
+    jsonRpcProvider({
+      rpc: (chain) => ({
+        http: chain['fallbackRpc'],
+      }),
+    }),
+    publicProvider(),
   ]
 )
 
@@ -102,7 +111,13 @@ const { wallets } = getDefaultWallets({
   chains,
 })
 
-const connectors = connectorsForWallets([...wallets])
+const connectors = connectorsForWallets([
+  ...wallets,
+  {
+    groupName: 'Other',
+    wallets: [rabbyWallet({ chains })],
+  },
+])
 
 export const wagmiConfig = createConfig({
   autoConnect: true,